refactor(priority)!: Remove update item priorities
Remove the support for updating download client item priorities.  This should be handled
at the time items are grabbed and added to the download client so that users may
manually adjust priorities on a per-item basis after they've been added.  Having this
tool come in periodically and change all priorities back proved to be more disruptive
than useful.  If there turns out to be need for something like this going forward, it
should be as a separate sub-command that can be run manually as needed or used as a
custom Servarr script.