refactor(servarr): Improve Servarr config lookup
Improve the handling of Prunerr config and the download client settings looked up from
the Servarr API.  Mostly this is about inverting the relationship between the config and
the clients.  Put the transmission/Servarr API clients into the config dictionaries
rather than attaching the config dictionaries to the clients as attributes.