feat(sync): Command sync Servarr state w/ download
Add a sub-command that synchronizes download client state with Servarr
state, primarily via download item location directories in the
download client.  Also create symbolic links for imported files to
make it easier to introprospect individual files.

This is also the foundation for use as a Sevarr custom scripts to
handle Servarr workflow events.