fix(history): Collect history per series/movie
We still need to be able to page through history for the `$ prunerr sync` sub-command so
this largely reverts the previous commit and adjusts it so we can preserve that hsitory
search for the rare case but match history on a per-movie/series/etc. basis for the
common case of responding to a Servarr event via Custom Sctipts.

This should also resolve the issue of getting fresh history for long running tasks such
as `$ prunerr daemon`.

This reverts commit 3fdd8597d9474ba91be19e88fff009d4019bb2a2.