fix(imported): Use per-item Pruner JSON data file
The custom script for the imported Servarr event seems to need to exit before it
proceeds with any further importing.  Detaching a process and exiting and anything else
I tried didn't seem to address the issue of the custom script moving the downloaded
files out from under Servarr before it can import the rest of the files in a multi-file
download item.

We already have a long-running process, `$ prunerr daemon`, so use that to separate the
custom script's exit from the moving of the download item after import has completed.

While we're at it, I've been thinking of using a per-item data file anyways and that has
a number of advantages.   It can be used to track the time since the last file imported
by Servarr for a given item, which solves the imported event async issue.  But it also
makes keeping download item state in sync much easier:

1. Update state in the data file
2. Use timestamps in the data file to find unhandled events
3. Handle those events and add timestamps