feat(sync): Do away with custom script handling
At least for Servarr Custom Scripts, we have to rely on the `$ prunerr daemon` loop to
followup and do the real processing because we can't count on the necessary history
record data being available from the Servarr API at the time the custom script is
executed.  It may be that this isn't an issue for Servarr WebHooks, that we can return
the response to the WebHook request early, or some other way of dealing with similar
issues under WebHooks, but it just isn't worth it.

We already need to poll the download client anyways (see comments for why), and having
in the `$ prunerr sync` sub-command all the handling necessary to synchronize download
client state with Servarr state makes the whole system more robust and greatly
simplifies the code.  Do away with Servarr custom script handling all together along
with any aspirations to switch to WebHooks.

This also allows us to switch to our own, isolated container without any further changes
to Prunerr.