feat(docker): Container status service crash
When using an `s6-overlay` base image, such as `lsiobase/*`, a crash of the main service
process, `$ prunerr daemon` in this case, doesn't cause the container to exit.  IOW, the
container's status doesn't reflect the actual status of the service the container in
responsible for.

It's possible to duplicate host filesystem permissions inside the container with `$
docker run --user ... --add-group ...`, and thus also in `./docker-compose.yml`.  As such we
don't need to use a `lsiobase/*` base image to accomplish that.  Better to use the
official `python` base image if we don't need anything else provided by `s6-overlay`.

This reverts commit 6d927868324cf3793077f1f88bf1eb341d5ed848.