feat(dist)!: Convert setup.py -> pyproject.toml
Having revisited `./pyproject.toml` it seems all the tooling I currently depend on
finally supports it when using recent `$ pip` versions. RIP `./setup.py`.

The only thing that I had to workaround was tox's `usedevelop = true`, but I would argue
I shouldn't have been using that by default anyways.  It's best to test against an
actual `$ pip install ...` of the actual built distribution which is what tox does
without `usedevelop = true`.  It's pretty easy to install the distribution in editable
mode when one needs to debug changes live in "real-time" as is captured in the
`./Makefile`.