build(make): Clarify real vs PHONY targets
For a while now I've been playing with adding more context clues, in this case using
`./` to clearly indicate that what follows is a repo-relative real filesystem path and
using a trailing `/` to clearly indicate that what precedes is a real filesystem path to
a directory.  I've found it to make visually scanning things like `./Makefile`s much
clearer and easier.   Neither have I found there to be any issues, so it's time to make
the change.