build(devel): Disambiguate development extra name
I've never liked using `dev` as a shorthand for development.  In *NIX land that means
"device" and it could be a prefix for any other number of words.  Whereas with just 2
more letters `devel` is unambiguous and matches the Debian package naming convention.
And if you're going to back anyone for well thought out conventions, it's Debian.