fix(docker): Propagate env vars in compose config
Any variable substitution in the `./docker-compose*.yml` files, such as `${PUID:-1000}`,
is explicitly intended to be taken from the `./.env` file, but the `./Makefile` wasn't
even rendering the `./.env.in` template.  I can't recall but I assume I just copied and
pasted the `./.env.in` as the target pre-requisite and just forgot the delete the `.in`.

Fix the pre-requisite and ensure that all variables expanded in the
`./docker-compose*.yml` files are present in the rendered `./.env` file.