build(docker): More efficient build cache usage
The longest step of the build process is installing the Python requirements, so make
that step dependent on just the `./requirements*.txt` file to avoid having to rebuild it
when anything changes in the checkout.  Similarly, installing this package in
develop/editable mode is the step that has the most churn so it should be last.