fix(history): Narrow history search by type, WIP
I'm kicking myself for only checking now, but I was looking at the source to see if
there might be a way to request multiple event types from the global history endpoint.
While looking at that I saw that the endpoint accepts an undocumented `downloadId`
parameter.  This is true for both Sonarr and Radarr and I tested it in the
debugger. This is obviously the way to go and removes the need for most of the code I've
written to search global history.  So that's the next step.