feat(move)!: Vastly more simple sync replacement
Total face-palm.  All we need to do to tell if Servarr has acted on a download item is
just to check if it's in the Servarr instance's download client's `Download` directory
settings and that it's not in that Servarr instance's queue.  Use just that to "sync"
the items location in the download client after the item has been acted on by that
Servarr instance.  With that change, it's much less about synchronizing Servarr state
with the download item, hence the change in terminology from `sync` back to `move`.