test(vcs): Commit hook missing dependency changes
I added a dependency to `./setup.cfg` and added a corresponding command to `./tox.ini`
but the commit hook just running `$ tox` seems not to have updated `./.tox/*/`:

    lint run-test: commands[3] | twine check
    ERROR: InvocationError for command could not find executable twine

That ended up being the wrong way to use `$ twine check` but the issue it exposes
remains.  Address this by using the `./Makefile` to ensure everything is up to date in
the commit hook.