build(docker): Minimize end user image size
I didn't like that the end user image includes a copy of the checkout just to install
the Python distribution/package from the checkout into the system Python.  Once
installed the checkout is no longer needed and shouldn't be a layer in the image.
Instead use a bind mount of the checkout only during the install.

I've been going back and forth on how the Python dist should be installed in the end
user image.  One could argue that it should be installed from an actual built wheel, but
I would argue that the wheel/sdist should be built in a container to minimize pollution
from the developer's host OS.  I don't think such a nested round trip is worth it
especially since the real answer to this kind of issue is to build actual release
artifacts, including container images, in CI.  So this will do for a local story for
now.