build(docker): Broken run-time ~foo/ passwd DB
I ran into a very inscrutable issue with `$ tox` failing only inside the development
container.  Even with the `${HOME}` environment variable set, it was creating a
directory in the checkout actually named `./~/`.  I confirmed that adding a real user to
the container resolved the issue with `$ tox`.

I dislike the pattern of using an `./entrypoint.sh` script as the container image's
`ENTRYPOINT`.  I like the elegance of the container's intended primary command being the
`ENTRYPOINT` directly.  Unfortunately, there are also many container use cases that
require aligning UID/GID inside and outside the container, mostly for file ownership
permissions issues.  As such, being able to change the UID/GID at run-time using
Docker's `--user` option is a hard requirement.  The only robust way to have a real user
in the `/etc/passwd` DB with the right UID/GID is to add that user to the container's OS
at run-time, so we need an `ENTRYPOINT` script to do that.

If the `ENTRYPOINT` script adds a user to the OS inside the container, then it must be
run as `root` which in turn means we can't use the `--user` option at all.  So we have
to `# su ...` to the run-time user at the end of the `ENTRYPOINT` script.  I really wish
Docker had a better answer for this use-case.