build(release): Move publishing outside container
It's not only easier to implement publishing release artifacts outside the container,
for access to credentials, but it also seems more correct now that I think about it.
Creating release artifacts makes sense to do in a container for consistency and
reproducibility but publishing to a remote is a user task and should be run as the user
and in the user's environment.