ci(release): Reduce CI platform dependency
Keeping as much consistent between CI and local development serves several purposes.  It
makes issues exposed on CI easier to understand and makes them easier to reproduce and
resolve locally.  Of course adding or switching CI platforms also becomes much simpler.
The only things we should depend on that is specific to a CI platform are services
provided by that platform that make runs faster or that give developers a better UX when
examining CI results.