build(deps): Improve host OS dependency installs
As far as I remember, the our only host OS requirements are `$ tox` and `$ docker` for
that branch.  Since `$ tox` is a Python dependency, we can install it in the
`./Makefile` without introducing a bunch of environment assumptions.  So add a target
build artifact for that install and explicitly add that as a prerequisite of all target
recipes that invoke `$ tox`.

It also seems like I missed when the `./var/log/init-setup.log` target
became meaningless, so clean that up.