ci(release): Missing private signing subkey
CI results showed and local testing confirmed that exporting a private key, importing it
into a temporary directory, removing its passphrase, exporting it again (now
unprotected), then importing *that* into *yet another* temporary directory and trying to
sign a file results in:

    $ gpg --homedir /tmp/gnupd.XXXXXXXXXX.d --local-user $(GPG_SIGNING_KEYID)! --sign ./foo.txt
    gpg: skipped "$(GPG_SIGNING_KEYID)!": No secret key
    gpg: signing failed: No secret key

So I assume it's not possible to import an unencrypted private key and use it as a
private key, or at least I'm done fighting `$ gpg` and resigning myself to adding a
passphrase to CI secrets.  FWIW, there's no difference between having the passphrase for
a private key in the same CI secrets space and having a user's passphrase in a clear
text file inside their `~/.gnupg` directory.