ci(release): New unlock key passphrase approach
I couldn't reproduce the last CI issue locally:

    + pgrep -la gpg-agent
    8128 gpg-agent --daemon --allow-preset-passphrase --max-cache-ttl 3153600000
    + /usr/lib/gnupg2/gpg-preset-passphrase --preset 2EFF7CCE6828E359
    + printenv GPG_PASSPHRASE
    gpg-preset-passphrase: caching passphrase failed: Not implemented
    make[1]: Leaving directory '/home/runner/work/python-project-structure/python-project-structure'

But I couldn't get it to work either.  the `$ gpg-preset-passphrase` command succeeded
but signing something still prompted for the passphrase.  OTOH, doing one signing
operation using `--passphrase-file` resulted in a second signing operation that didn't
prompt for the passphrase so try to use that approach on CI.