build(release): Fix versions derived from VCS
When `$ semantic-release` found no commits requiring a new release, the conditionals in
the `./Makefile` handling that condition only handled the first occurrence but not the
second occurrence that actually bumps the version and creates the tag.  Hence the
subsequent distribution was built with a `*.dev*` version.  Extend that logic to the
second occurrence to fix that.

Another issue was that the generated changelogs or release notes were using `*.dev*`
versions as well because we need to generate the commit that updates release notes
before we bump the version because the tag needs to be placed on that very commit.  Fix
this by explicitly passing in the version when building a release.

Having run into all this, I'm thinking we should just listen to `$ semantic-release` and
not publish releases when it says we shouldn't.