build(release): Unnecessary build local testing
I've been feeling conflicted about having any prerequisites be a part of building the
actual distribution packages.  On the one hand, it should indicate an issue if something
needs updating between the build process and the release process.  OTOH, the release
process should publish only what was built during the build process.  The `build-local`
prerequisite was causing the `./requirements*.txt` targets to be updated inside the
container when testing releases locally.  I can find no CI run that had the same issue,
so that pushes me over the edge to just removing the prerequisite.