build(release): Artifacts built with old version
When building on a branch that should trigger a release, we build the distribution's
installable packages before the release process where the version is bumped.  This means
we're not testing the actual code that will be released.  Move version bump and
changelog assembly from release to a pre-build phase to address this.