build(release): Fix version released from master
IMO, it should be the common case, if not enforced, that a final release is the same as
the last pre-release.  IOW, a final release is just the maintainer saying, "Yup, that
lase pre-release is good to go and is now the final release".  Unfortunately, `$
semantic-release version` *only* bumps versions if it finds commits since the last
release and so we can't use it directly to bless the last pre-release as the next final
release.  Also unfortunately, the only way it gives us to force the next version is to
specify whether the next version is a major, minor, or patch version.  So working around
these limitations isn't very elegant and I may end up switching back to `commitizen` but
for now this what I had to do to extract the type of release to force `$
semantic-release version` to bump the version.