build(vcs): Change default VCS host to GitLab
I've been wanting to make this switch for a while.  Long enough that I can't even
remember which incident it was when GitHub first took on an evil smell, but that's
become a trend.  There are a lot of technical reasons I prefer GitLab even without the
ethical concerns.  Even without those technical reasons, it's usually a good idea to
support the 2nd biggest player in a field just to keep ecosystems healthy.