ci(github): Missing secret variables
Checking for the presence of secret variables as the conditional for running parts of
the release process resulted in hiding some CI failures.  Make those conditionals more
explicit on being run in a CI environment to avoid silent failures.  Also add GitHub
secret variables that obviously should be in place.