ci(release): Debug broken GitLab push remote URL
I'm very confused why this isn't working.  I've confirmed that the `$ sed` expression
works locally and that I can push to the resulting remote URL in a clean container and
that I can't push to the original URL in that same clean container.  So my best guess is
that maybe the remote URL is preserved between runs so that my previous attempts have
screwed the remote URL up beyond all recognition.  But that doesn't make sense from the
CI/CD job output that clearly indicates creating a fresh git repo.  I'm at a loss, so
lets just try this.