build(docker): Faster, smaller image, multi-stage
Build an installable wheel package in one build stage then copy it into the final image
to be installed.  This removes the bind mount from the final image and IIUC this
improves both the size of the final image and the efficiency of the build cache.

It also achieves a goal I've been wishing for, have the container built by installing
the same package that is published in releases.  This approach allows us to both build
the package in a clean, reproducible container and then install it into the final
container image.