build(docker): Improve build cache efficiency
Using the end-user image as a base for the development image means caching is *always*
busted for the development image because the first layer is always changed.  Abandon
that approach and duplicate the required bits of the end-user image in the development
image to maximize cache efficiency.  I hate that this violates DRY, but the
`./Dockerfile*`s should change infrequently in the common use case so the violation is
worth it to save developer time in the inner loop of the development cycle.