build(docker): Generate release notes in container
We need `$ towncrier` to create the release commit and it needs to be on the same
`sys.path` as the package being developed.  Unfortunately this means we need the
container to create the release notes we need to bake into the container and that in
turn means we need to build the container twice when the version is bumped.