build(release): Support local testing, delete tag
When testing a release locally on a branch whose latest tag is out of date compared with
`master`, it's necessary to delete the tag for the "next" release locally.  The fetch
intended to ensure `$ towncrier` can compare commits, however, was restoring any such
deleted tag.  Fetch only the `--compare-with` branch so that we can still test releases
locally.