build(tox): Optimize development task times
Though [the tox
documentation](https://tox.wiki/en/latest/user_guide.html#system-overview) explicitly
says that `$ pip install` won't upgrade a dependency if it already exists in the `venv`,
I tested changing a version in `./requirements-devel.txt` and `$ tox` correctly detected
the change and recreated the `venv`.  I'm guessing that paragraph in the docs is out of
date.  [Later in the docs](https://tox.wiki/en/latest/user_guide.html#cli), it more
explicitly says it *does* try to detect when to recreate `venv` and recommends the
`--recreate` option as a kind of last resort if strange behavior is observed.  Given
that we pin our versions in `./requirements-devel.txt` and the `./Makefile` should
update that file when needed and that tox `venv` recreation is most of the build time
when it happens, better to trust tox's recreation detection by default to save developer
time in the inner loop of development.

Also trim unnecessary prerequisites to reduce unnecessary build times.