build(docker):: Redundant requirements install
Since the canonical set of test, lint, and check commands is defined in `./tox.ini`, we
should consider `$ tox` to be the primary tool to run such tests and checks for
developers.  Indeed it's already the default command of the development container.
Since installing the requirements is the most time-consuming part of the image build
process, lets install the development requirements only once in the `./.tox/py3##/`
Python environment.