build(release): Frozen build reqs for supported py
Build frozen/pinned versions of the build and release tools for every supported Python
version.

I realized that we built per-python-version requirements files for the host install
dependencies but not for the build and release tools dependencies.  Both of these are
only used on the host and aren't actually used in the per-python-version virtual
environments so they should have the same behavior.  The question of which way to go on
this matter for both files is a trade off between longer build times when versions are
updated/upgraded vs allowing developers to develop from a wider range of Python
versions.  I ended up opting for wider development environment support because that is
very likely to be relevant for real projects and updating versions is infrequent.