build(py): Build tools venv wrong reqs file path
When I extended per-python-version requirements compiling to the `./.tox/build/` Python
environment, I forgot to make the corresponding change in `./tox.ini`.  The only
variable set by tox itself that contains the Python version is `base_python` and that's
the absolute executable path.  Unfortunately, the support for variable expansion in
`./tox.ini` is minimal and there's certainly no support for paring `/usr/bin/python3.11`
down to `python3.11` let alone down to `py311`.  So that means this change requires
passing an environment variable into `$ tox -e build`.

I decided to match the behavior of `$ tox` for determining the default `base_python`
value and added that logic to `./Makefile`.  An earlier approach of choosing the latest
installed Python seems like a good fallback, so I kept it as well.