build(make): Avoid redundant target updates
For real targets, we should limit the prerequisites of the target to only those for
which changes actually require the target to be rebuilt/updated.  Anything required by
the target's recipe but that changes to don't affect the built target should be built in
a sub-`$(MAKE)` instead.

This is as opposed to `.PHONY` targets that always run regardless of their
prerequisites.  IOW, the prerequisites of `.PHONY` targets are a matter of convenience.