build(docker): Parallel test run packaging clash
Parallel runs of the tests in all the Docker image Python version variants were failing
intermittently, mostly with errors about missing files while `$ tox` was trying to build
the package.  Unfortunately, it seems like the current Python build tools don't
provide *any* way to control where it creates the staging directory from which it
creates the archive.  It's *always* done in `./project-name-#.#.#/`.  Thus when multiple
containers are run at once, they're all building their package in the same directory and
stomp on each other.

Outside of containers for local development, when `$ tox` runs in parallel, it first
creates one package to be used by all the test envs.  Building the package in each
container is thus a waste of time as well, so mimic `$ tox` and first create just one
package and share it between all the containers.