ci(github): Workaround permissions error
This error happened in the `build-test` job on GitHub Actions after the `bump` job
succeeded.  Then I re-ran the `bump` job and it failed in the same way.  I can't think
of anything between `bump` and `build-test` that would have created that directory
except the caches.  Of course neither can I think of a way that `./var/log` could have
made it into the cache with thew wrong owner.  So clearing the caches is a shot in the
dark:

    + mkdir -pv var/log/
    mkdir: can't create directory 'var/log/': Permission denied
    make: *** [Makefile:757: var/log/host-install.log] Error 1