build(release): VCS remote in sync w/ published
Initially I thought the VCS remote should only reflect a release after it has been
successfully published.  In reality, however, there's too much that can go wrong in the
integrations with the various services involved in publishing releases.  What if a
built package is successfully pushed but then pushing a container image fails?  What if
everything that actually publishes release artifacts works but then updating the docs
fails?  Better that VCS reflects the attempt and not everything is published than
something is published but VCS shrugs its shoulders when asked.