fix(docker): Workaround other service net diconn
If the `transmission` service uses `network_mode: "service:..."`, such as a VPN, then it
may become inaccessible on any IP other than `localhost` when the service whose network
it uses is restarted.  As such, if the `transmission` service should be accessible from
it's IP on the LAN set the `LAN_IP=...` in the `./.env` so that the `HEALTHCHECK` fails
and the container is restart when that happens.