build(release): Enforce release notes on push
Originally, I thought that contributors should be able to push freely to their feature
branches and as such we shouldn't enforce the presence of release notes in the VCS push
hook.  But CI for feature branches *should* fail if release notes are missing and having
CI and the VCS hooks have different behavior is more surprising than having the VCS hook
fail when just trying to "back up" work in progress to the remote.  Developers can push
freely using `--no-verify` in that case.