build(py): Default latest py installed on host
We already have support in the `./Makefile` for defaulting to the latest Python version
installed on the host.  I have made arguments to myself for both that approach and for
the approach of defaulting to the latest Python version supported by the project.  Until
I can come up with a clear argument for a winner, I'll stick with existing logic.