build(docker): Consistent variable default values
Add defaults for all variables where it's ever possible for there to be a valid default
value, mostly non-secrets.  I've gone back and forth on whether the default for `PGID`
should be `100`, the conventional `GID` for the `users` group on *NIX systems. or
`1000`, the default private GID for the first/main user in most Linux distros.  I'm
still not sure which is correct but either way, better not to have a mix.