build(docker): Wrong container exit code
I'm assuming the intention of the `/etc/s6-overlay/s6-rc.d/svc-transmission/finish`
script from the container image is to ensure the daemon shuts down even if it somehow
detached from the `# s6-notifyoncheck ...`.  I guess the idea is that the image wants `$
transmission-daemon` to actually shutdown before s6 restarts it via the service.

In our case, we prefer the container exit when the
`/etc/s6-overlay/s6-rc.d/svc-transmission` service exits for any reason so that the
incident is visible in the container management tools, that's the reason for modifying
the `/etc/s6-overlay/s6-rc.d/svc-transmission/finish` script in the first place.  So
it's probably better to rely on s6 `# init` to gracefully terminate any detached daemons
as a part of shutdown than to use `$ transmission-remote --exit` as some sort of double
check.  Especially given that if `$ transmission-daemon` is hung for some reason, then
`$ transmission-remote --exit` will cause the `finish` script to fail before it has a
chance to tell the container to shutdown.