build(deps): Workaround pip HTTP file cache race
The workaround here is to repeat some number of times the commands which are subject to
the race condition, running pip in parallel, in the hopes that subsequent runs of only
the failed targets will succeed.  This is a less correct workaround but one I suspect is
better pragmatically because I assume the caching saves more time even given the re-runs
than disabling the cache would save by avoiding re-runs.  I'll test that next.  Here's
the time for this pragmatic workaround:

    $ time (make upgrade; make upgrade; make upgrade)
    ...
    real    1m46.404s
    user    16m9.194s
    sys     1m19.079s