build(deps): Pip race, separate HTTP file caches
This workaround I would say is the most correct but it doesn't seem to work.  I still
see the `WARNING: Skipping page https://pypi.org/simple/wheel/ because the GET request
got Content-Type : .The only supported Content-Type is text/html` race condition.  I
confirmed that the `~.cache/pip-tools-*-*/` directories are created by `$ pip-compile`
so the option is taking effect but all the created directories are empty so somehow this
option isn't affecting the actual caching and this doesn't workaround the race
condition.

This reverts commit ff563e773521c29c9498cf476a2a17335b47e285.