build(deps): Pip race, disable HTTP file cache
This would be the least optimal workaround but it doesn't work, I still see the
`WARNING: Skipping page https://pypi.org/simple/wheel/ because the GET request got
Content-Type : . The only supported Content-Type is text/html` race condition errors.
So it turns out the pragmatic, not entirely correct, workaround is the only one.  Back
to that next.