ci(docker): Workaround interactive output w/o TTY
In at least 2 cases I've seen repetitive, noisy interactive Docker image pull output in
CI:

    ...
    aa54add66b3a Downloading [>                                                  ]  540.7kB/54.59MB
    aa54add66b3a Downloading [===>                                               ]  4.295MB/54.59MB
    ...
    1e4aec178e08 Extracting [======================>                            ]  25.07MB/55.05MB
    1e4aec178e08 Extracting [=========================>                         ]  27.85MB/55.05MB
    ...

The 2 cases I've identified are for [`$ docker compose create
...`](https://gitlab.com/rpatterson/python-project-structure/-/jobs/3797287139#L97) and
[`$ docker compose run
...`](https://gitlab.com/rpatterson/python-project-structure/-/jobs/3797287139#L564)
when the image hasn't been previously built, such as in the `$ make upgrade-branch`
scheduled pipeline.

I was able to [simulate this locally](https://superuser.com/a/1430883/96351) using
`$ (setsid ...) </dev/null |& cat` and confirm that this also happens with `$ docker
buildx build --pull ...` and with `$ docker compose pull ...`.  In fact, the only case I
could find where pulling an image properly detects whether the session is interactive is
`$ docker pull` so it seems we shouldn't rely on any part of the Docker tooling to
handle this.