lint(perf): Optimize Python literal usage
I have always liked using `dict(foo="bar")` where I was using the dictionary to assemble
or manage kwargs to be passed into a function call because the syntax matches the
function call.

PyLint notes that doing so introduces an extra function call and the attendant
performance overhead.  None of these instances are important for performance but better
to avoid the habit than add a bunch of inline ignores.