build(release): Don't bump version on PR branches
We shouldn't count on the version tags on commits made to release
branches (e.g. `develop` and `master`) being in the VCS history for feature/PR branches.
For that matter, the version-bump/release commits should never be on feature/PR branches
so it doesn't make sense to test against such a commit.  IOW, on feature/PR branches do
everything we can to check that we could bump the version and make the release commit
and tag without actually doing it.