build(make): Generalize sharing universal wheel
Within a given `$ tox ...` invocation, tox itself manages building and installing the
package efficiently.  For example, `$ tox run-parallel` builds the wheel once and then
installs it into each venv simultaneously.  Across `$ tox ...` invocations, however, tox
has no way of efficiently checking if an existing wheel is current and thus rebuilds the
wheel for each invocation unless explicitly given a package by `--installpkg`.

Add support for passing an up-to-date built wheel between `$ tox ...` invocations using
a symlink.  Not currently used on this `dist` branch but needed on the `docker` branch
and may be useful on other branches or this branch in the future.