build(release): Test before bump and release
Begin rethinking the release process with experience from working on CI/CD.

Previously, I've been working under the assumption that we should test exactly what we
release.  That means on branches that should publish a release, we bumped the version
and updated the changelog *before* testing and then published the release after if tests
passed.  In real-world CI/CD, this creates a number of hassles, mostly because most
CI/CD platforms expect to checkout only the commit that triggered the run.  For pipeline
jobs that run after the version bump, this means pulling that new commit and that just
breaks assumptions made by the CI/CD platform all over the place, much more than I
expected.  It also means we have to ignore the version bump commits in CI/CD and most
CI/CD platforms seem to expect pushing version tags to be used for the CD part of CI/CD.
IOW, this has us fighting CI/CD platforms.

I still think it's more ideal to release exactly what was tested and to test exactly
what will be released.  I can also argue, on the other hand, that if there's something
in the release process that breaks something that would be caught by the tests if
run *after* the release process, then fixing that is about fixing the release process
and not necessarily the job of CI/CD to reveal in normal operation.  So if the trade off
is between "the perfect" of releasing exactly what we test and a pragmatic choice not to
fight CI/CD, it seems like the latter is worth the compromise.  Further, I was able to
run into the GitLab free tier CI/CD minutes limit very quickly just by running a nightly
job to upgrade requirements, so I think we should similarly skip doing another test run
on the release commit.

In summary, change the process from `bump -> test -> release` to a separate `test`
pipeline followed by the also separate `bump -> release` pipeline.