ci(release): Separate release version tag pipeline
Switch from doing everything in one pipeline to separate `build-test` and `release`
pipelines.  If all Python versions build and pass tests and the branch and changes
require a release, then bump the version and push the tag.  The subsequent pipeline run
for the pushed tag will then perform the release.