ci(release): More secure release restrictions
Thinking about trusted vs untrusted inputs, it seems like we should consider what is
pushed to a get remote to be an untrusted input in terms of CI/CD and that we switch
from untrusted to trusted input via the security controls that project hosts provide us,
namely protected branches.  IOW, only stuff pushed to protected branches should be
considered trusted input and thus something we trust to trigger releases.