ci(release): Prevent duplicate pipelines
Unfortunately, GitLab CI/CD doesn't provide an elegant way to prevent duplicate pipeline
runs.  What I would like to be able to express through the CI/CD configuration is, "When
I push a protected tag formatted like a version, that commit is a release whose parent
has already been tested and now needs to be released."  GitLab only allows us to make a
conditional on the tag name for pipelines triggered by pushing a tag because the
`${CI_COMMIT_TAG}` is only available on such pipelines.  Pushing the same commit to the
branch on which that tag is placed, however, *also* triggers a separate pipeline where
`${CI_COMMIT_TAG}` is *not* available so we can't exclude that pipeline by matching the
tag name.

https://docs.gitlab.com/ee/ci/jobs/job_control.html#avoid-duplicate-pipelines