ci(release): Missing branch name on release tag
When we switched to publishing releases from the *tag* pipeline for the release version
tag, we encountered more issues because the only reference name GitLab CI/CD provides in
tag pipelines is the tag name and we have no access to the branch name in the GitLab
CI/CD configuration.  As such, the release job doesn't have access to any cache.  This
also causes issues with using the branch name in the release process, such as detecting
whether to publish a pre-release for `develop` vs `master`.

So switch back to releasing from the branch push pipeline.  This does have the added
benefit of being able to unify the version bump and release processes into one job and
saving more CI minutes in per-job overhead.

I'm starting to feel less positive about GitLab CI/CD now with all this hassle around
what triggers pipelines and duplicate pipelines.  In this case, I do think developers
want a tiny bit of magic ala GitHub Actions.  Since GitLab CI/CD *does* provide access
to the `$CI_PIPELINE_SOURCE` which gives us a way to discriminate on the pipeline
trigger, and easy solution would be to provide both `$CI_COMMIT_BRANCH` and
`$CI_COMMIT_TAG` regardless of what triggered the pipeline.