ci(release): More elegant version bump commit skip
I found trying to exclude the version bump commits in the job conditionals fragile in
GitLab CI/CD and not possible in GitHub Actions.  One could argue that if it's specific
commits that need to skip CI, then that's where the logic belongs.  The only benefit to
having it done in the conditional is that the pipelines then don't show up in the GitLab
CI/CD web UI but one could also argue it's valuable to see releases represented there.