build(vcs): More correct compare branch detection
Thinking about it further, using Git's upstream/merge-remote configuration isn't
necessarily correct for determining which reference should be used as the comparison for
determining conventional commits version bump and/or checking for required release
notes.  Users may use different upstream/merge-remotes and push remotes for various
reasons, best to let that be separate.  Determining what to use for comparison is a
matter of project release policy not developer configuration.

Fall back to more minimal and explicit detection from `./.git/config` and defining the
policy in the `./Makefile`.