build(py): Mismatched canonical Python version
I've gone back and forth in terms of deciding the canonical/official/default Python
version we support.  Initially I was using the latest version blessed as supported at
python.org.  Later, when adding a build host image in which to build the end-user
images, I started using the default python version used by the build host image.  The
latter has grown on me.  There aren't any specific issues either way, but it makes sense
to use the second to latest supported version just to avoid bleeding edge issues by
default.  It also makes sense to follow the lead of Linux distributions and when they
find it appropriate to upgrade Python versions just as a curator of issues projects may
encounter when upgrading.  So I'm making a call to got with the build host version as
the canonical version.