ci(gitlab): Yet another unknown status fix, take 3
Ugh, now that we're testing and publishing releases from the version tags,
the *coverage* badge displays `unknown`.  The kludge of having an always successful job
for whichever type of pipeline from which we're *not* publishing releases seems to be
working for the pipeline status so lets see if it works for both badges when we reverse
the pipeline source we release from.  IOW, lets switch to releasing from tags to
releasing from the release branches.